+++ src/game/model/player.cpp
@@ -3,18 +3,18 @@
 namespace model {
     
     Player::Player() {
-
+        
     };
 
     Player::~Player() {
-
+        
     };
 
     void Player::draw(Renderer * renderer) {
 
     };
 
-    void Player::update(Uint32 dt, const Uint8 * input) {
+    void Player::update(float dt, const Uint8 * input) {
 
     };
 }
