+++ src/game/scene/abstract_scene.cpp
@@ -7,7 +7,7 @@ namespace scene{
         for(gameObjIter it = this->objects.begin(); it != this->objects.end(); it++) it->second->draw(renderer);
     }
     
-    void AbstractScene::update(Uint32 dt, const Uint8 * input) {
+    void AbstractScene::update(float dt, const Uint8 * input) {
         for(gameObjIter it = this->objects.begin(); it != this->objects.end(); it++) it->second->update(dt, input);
     }
